third_party/blink/web_tests/http/tests/websocket/send-arraybuffer-expected.txt

WebSocket: Send ArrayBuffers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: Message #0.
PASS PASS: Message #1.
PASS PASS: Message #2.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
